+++ example/invalidate_path.c
@@ -19,13 +19,13 @@
  *
  * ## Compilation ##
  *
- *     gcc -Wall @file `pkg-config fuse3 --cflags --libs` -o invalidate_path
+ *     gcc -Wall invalidate_path.c `pkg-config fuse3 --cflags --libs` -o invalidate_path
  *
  * ## Source code ##
- * \include @file
+ * \include invalidate_path.c
  */
 
-#define FUSE_USE_VERSION 31
+#define FUSE_USE_VERSION 34
 
 #include <fuse.h>
 #include <fuse_lowlevel.h>  /* for fuse_cmdline_opts */
@@ -212,6 +212,7 @@ int main(int argc, char *argv[]) {
 	struct fuse_args args = FUSE_ARGS_INIT(argc, argv);
 	struct fuse *fuse;
 	struct fuse_cmdline_opts opts;
+	struct fuse_loop_config config;
 	int res;
 
 	/* Initialize the files */
@@ -271,8 +272,11 @@ int main(int argc, char *argv[]) {
 
 	if (opts.singlethread)
 		res = fuse_loop(fuse);
-	else
-		res = fuse_loop_mt(fuse, opts.clone_fd);
+	else {
+		config.clone_fd = opts.clone_fd;
+		config.max_idle_threads = opts.max_idle_threads;
+		res = fuse_loop_mt(fuse, &config);
+	}
 	if (res)
 		res = 1;