+++ test/travis-build.sh
@@ -25,7 +25,7 @@ chmod 0755 "${TEST_DIR}"
 cd "${TEST_DIR}"
 
 # Standard build
-for CC in gcc gcc-7 clang; do
+for CC in gcc gcc-7 gcc-10 clang; do
     mkdir build-${CC}; cd build-${CC}
     if [ "${CC}" == "clang" ]; then
         export CXX="clang++"
@@ -35,15 +35,13 @@ for CC in gcc gcc-7 clang; do
     else
         build_opts=''
     fi
-
-#    meson -D werror=true ${build_opts} "${SOURCE_DIR}" || (cat meson-logs/meson-log.txt; false)
-#    ninja
-    cmake -G "Unix Makefiles" \
-            -DOPTION_BUILD_UTILS=ON \
-            -DOPTION_BUILD_EXAMPLES=ON \
-            -DCMAKE_INSTALL_PREFIX="${SOURCE_DIR}"/install \
-            -DCMAKE_BUILD_TYPE=Debug "${SOURCE_DIR}"
-    make -j
+    if [ ${CC} == 'gcc-10' ]; then
+        build_opts='-Dc_args=-flto=auto'
+    else
+        build_opts=''
+    fi
+    meson -D werror=true ${build_opts} "${SOURCE_DIR}" || (cat meson-logs/meson-log.txt; false)
+    ninja
 
     sudo chown root:root util/fusermount3
     sudo chmod 4755 util/fusermount3
@@ -59,15 +57,9 @@ for san in undefined address; do
     mkdir build-${san}; cd build-${san}
     # b_lundef=false is required to work around clang
     # bug, cf. https://groups.google.com/forum/#!topic/mesonbuild/tgEdAXIIdC4
-    #     meson -D b_sanitize=${san} -D b_lundef=false -D werror=true "${SOURCE_DIR}" \
-    #       || (cat meson-logs/meson-log.txt; false)
-    # ninja
-    cmake -G "Unix Makefiles" \
-            -DOPTION_BUILD_UTILS=ON \
-            -DOPTION_BUILD_EXAMPLES=ON \
-            -DCMAKE_INSTALL_PREFIX=/home/<USER>/FUSE/install \
-            -DCMAKE_BUILD_TYPE=Debug "${SOURCE_DIR}"
-    make -j
+    meson -D b_sanitize=${san} -D b_lundef=false -D werror=true "${SOURCE_DIR}" \
+           || (cat meson-logs/meson-log.txt; false)
+    ninja
 
     # Test as root and regular user
     sudo ${TEST_CMD}