+++ ChangeLog.rst
@@ -1,3 +1,44 @@
+UNRELEASED CHANGES
+==================
+
+* <please document user-visible changes here when committing them>
+
+libfuse 3.10.3 (2021-04-12)
+===========================
+
+* Fix returning d_ino and d_type from readdir(3) in non-plus mode
+  
+libfuse 3.10.2 (2021-02-05)
+===========================
+
+* Allow "nonempty" as a mount option, for backwards compatibility with fusermount 2. The
+  option has no effect since mounting over non-empty directories is allowed by default.
+* Fix returning inode numbers from readdir() in offset==0 mode.
+* FUSE filesystems can now be mounted underneath EXFAT mountpoints.
+* Various minor bugfixes.  
+
+libfuse 3.10.1 (2020-12-07)
+===========================
+
+* Various minor bugfixes.
+
+libfuse 3.10.0 (2020-10-09)
+===========================
+
+* Add FUSE_CAP_CACHE_SYMLINKS: allow caching symlinks in kernel page cache.
+* Various minor bugfixes and improvements.  
+
+libfuse 3.9.4 (2020-08-09)
+==========================
+
+This was an "accidental" release, it is equivalent to 3.9.3.
+
+libfuse 3.9.3 (2020-08-09)
+==========================
+
+* Fixed compilation under OS X and µClibc.
+* Minor bugfixes and doc updates.
+
 libfuse 3.9.2 (2020-06-12)
 ==========================